# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wonder Toys Sheets API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://YOUR_PUBLIC_HOSTNAME"
    }
  ],
  "paths": {
    "/append": {
      "post": {
        "operationId": "appendRow",
        "summary": "Append a vending record to the sheet",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "location",
                  "product"
                ],
                "properties": {
                  "location": {
                    "type": "string"
                  },
                  "product": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "nullable": true
                  },
                  "note": {
                    "type": "string",
                    "nullable": true
                  },
                  "user": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appended row"
          },
          "401": {
            "description": "Invalid API key"
          }
        }
      }
    },
    "/last-product": {
      "get": {
        "operationId": "getLastProduct",
        "summary": "Get the last product added for a location",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest product info"
          },
          "404": {
            "description": "No data for location"
          }
        }
      }
    },
    "/search": {
      "post": {
        "operationId": "searchRows",
        "summary": "Search rows with optional filters",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {
                    "type": "string"
                  },
                  "product": {
                    "type": "string"
                  },
                  "since_ts": {
                    "type": "number"
                  },
                  "until_ts": {
                    "type": "number"
                  },
                  "limit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 500,
                    "default": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matching rows"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}